move and improve component for a single board
